Add files via upload
This Python notebook contains the plots for possible Helix binary systems with M,L dwarf companions. These plots aim to show whether these dwarfs can be a possible companion in the Helix. This applies to companions at any orbital period. This notebook also delineates the supplementary aim of this project.